+++ Makefile
@@ -33,8 +33,8 @@ SRCS = cmd_cmd_invocation.c cmd_cmd_invocation2.c cmd_exec_command.c		\
 	path.c g_cwd.c string_node2string.c env_expander.c heredoc_expander.c	\
 	expand_string_node.c split_expanded_str.c t_var.c t_var2.c				\
 																			\
-	builtin.c builtin_echo.c builtin_env.c builtin_exit.c builtin_cd.c		\
-	builtin_cd_path.c builtin_cd_chdir.c builtin_cd_cdpath.c				\
+	builtin.c builtin_fd_list.c builtin_echo.c builtin_env.c builtin_exit.c	\
+	builtin_cd.c builtin_cd_path.c builtin_cd_chdir.c builtin_cd_cdpath.c	\
 	builtin_export.c builtin_export2.c builtin_pwd.c builtin_unset.c		\
 																			\
 	str_utils.c shell_initialization.c minishell.c	interactive_shell.c
