+++ cmd_exec_builtin.c
@@ -10,47 +10,6 @@
 #include "builtin.h"
 #include "minishell.h"
 
-t_fd_list	*fd_list_add_fd(t_fd_list **lst, int fd)
-{
-	t_fd_list	*new_lst;
-	t_fd_list	*tmp;
-
-	if (!lst)
-		return (NULL);
-	new_lst = ft_calloc(1, sizeof(t_fd_list));
-	new_lst->fd = fd;
-	if (!new_lst)
-		return (NULL);
-	if (!*lst)
-		*lst = new_lst;
-	else
-	{
-		tmp = *lst;
-		while (tmp->next)
-			tmp = tmp->next;
-		tmp->next = new_lst;
-	}
-	return (new_lst);
-}
-
-void	fd_list_close(t_fd_list **lst)
-{
-	t_fd_list	*current;
-	t_fd_list	*tmp;
-
-	if (!lst)
-		return ;
-	current = *lst;
-	while (current)
-	{
-		close(current->fd);
-		tmp = current;
-		current = current->next;
-		free(tmp);
-	}
-	*lst = NULL;
-}
-
 /* Configure input redirection for builtin command.
  *
  * command: Command.
@@ -81,7 +40,7 @@ static int	builtin_set_in_red(t_command_invocation *command,
 				*stdinfd = dup(*stdinfd);
 			if (red->fd == *stdoutfd)
 				*stdoutfd = dup(*stdoutfd);
-			if  (dup2(fd, red->fd) == ERROR || close(fd) == ERROR
+			if (dup2(fd, red->fd) == ERROR || close(fd) == ERROR
 				|| !fd_list_add_fd(fd_list, red->fd))
 				return (put_redir_errmsg_and_ret(ERROR,
 						red->fd, strerror(errno)));
